moved task list diff code to task package
